Move passes() into new pass_matrix.py, add build_pass_matrix()
Separates the fixed-delta pass/fail matrix (used by analyses 1 and 3) from
committee_pool.py, which now only builds the raw long-format scores CSV.
pass_matrix.py holds passes() (moved, unchanged) plus the new
build_pass_matrix(), and produces pass_matrix.csv (56x10 dataset x model
boolean, NaN where a model has no data for a dataset).

Verified: matrix is byte-identical to the previously-verified version, and
reproduces the exact 23-accepted ground truth.
