+++ multabench/leaderboard/results/analysis_curation_sensitivity/pass_matrix.csv
@@ -1,0 +1,57 @@
+dataset,LightGBM,CatBoost,TabM,TabPFNv2,TabPFN-2.5,RandomForest,RealMLP,TabDPT,TabICLv2,XGBoost
+BIN_TEXT_FINANCIAL_CONSUMER_COMPLAINT,True,True,True,True,False,False,False,False,True,False
+BIN_TEXT_PROFESSIONAL_FAKE_JOB_POSTING,True,True,False,True,False,True,True,False,False,False
+BIN_TEXT_PROFESSIONAL_KICKSTARTER_FUNDING,True,True,True,True,True,True,True,True,True,True
+BIN_TEXT_SOCIAL_IMDB_GENRE_PREDICTION,False,False,False,True,False,False,False,False,False,False
+BIN_TEXT_SOCIAL_JIGSAW_TOXICITY,True,True,True,True,True,True,True,False,True,False
+BIN_TEXT_TRANSPORTATION_OSHA_ACCIDENT_INJURY_DATA,True,False,True,False,False,False,True,True,False,False
+MUL_TEXT_CONSUMER_PRODUCT_SENTIMENT,True,True,True,True,True,True,False,True,True,True
+MUL_TEXT_CONSUMER_WOMEN_ECOMMERCE_CLOTHING_REVIEW,True,True,True,True,True,True,True,True,True,True
+MUL_TEXT_FOOD_MICHELIN_GUIDE_RESTAURANTS,True,True,True,True,True,True,True,False,True,True
+MUL_TEXT_FOOD_WINE_REVIEW,True,True,True,,,True,True,True,True,True
+MUL_TEXT_FOOD_YELP_REVIEWS,False,False,False,False,False,False,False,False,True,False
+MUL_TEXT_HOUSES_MELBOURNE_AIRBNB,True,False,False,False,False,False,False,False,False,False
+MUL_TEXT_PROFESSIONAL_DATA_SCIENTIST_SALARY,False,True,False,True,True,True,True,True,True,True
+MUL_TEXT_SOCIAL_GOOGLE_QA_TYPE_REASON,False,False,False,True,True,False,False,False,False,False
+MUL_TEXT_SOCIAL_HEARTHSTONE_CARD_GAME_WARCRAFT,True,True,False,True,True,True,True,True,False,False
+MUL_TEXT_SOCIAL_NEWS_CHANNEL_CATEGORY,True,True,True,True,True,False,True,False,False,False
+MUL_TEXT_SOCIAL_SPOTIFY_GENRES,True,True,True,,,True,True,True,False,True
+MUL_TEXT_TRANSPORTATION_US_ACCIDENTS_MARCH23,True,True,False,True,True,True,True,True,True,True
+REG_TEXT_CONSUMER_AMERICAN_EAGLE_PRICES,False,True,True,False,False,False,True,False,False,True
+REG_TEXT_CONSUMER_BABIES_R_US_PRICES,True,True,True,True,True,True,True,True,True,True
+REG_TEXT_CONSUMER_BIKE_PRICE_BIKEWALE,False,False,False,True,False,False,False,True,True,False
+REG_TEXT_CONSUMER_BOOK_PRICE_PREDICTION,True,True,True,True,False,True,False,True,False,True
+REG_TEXT_CONSUMER_CAR_PRICE_CARDEKHO,False,False,False,False,False,False,False,False,False,True
+REG_TEXT_CONSUMER_JC_PENNEY_PRODUCT_PRICE,True,False,False,True,False,False,False,True,False,False
+REG_TEXT_CONSUMER_LAPTOP_INDIAN_PRICES,False,False,False,False,False,False,False,True,False,False
+REG_TEXT_CONSUMER_MERCARI_ONLINE_MARKETPLACE,True,True,True,True,False,True,True,True,True,True
+REG_TEXT_FOOD_ALCOHOL_WIKILIQ_PRICES,False,False,True,True,False,False,False,True,False,True
+REG_TEXT_FOOD_BEER_RATINGS,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_FOOD_CHOCOLATE_BAR_RATINGS,False,True,True,False,False,True,False,False,False,True
+REG_TEXT_FOOD_COFFEE_REVIEW,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_FOOD_RAMEN_RATINGS_2022,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_FOOD_WINE_POLISH_MARKET_PRICES,False,True,False,False,False,True,False,True,False,False
+REG_TEXT_FOOD_WINE_VIVINO_SPAIN,True,False,True,False,False,True,True,True,False,False
+REG_TEXT_FOOD_ZOMATO_RESTAURANTS,True,True,True,True,False,True,True,True,True,True
+REG_TEXT_HOUSES_AIRBNB_SEATTLE,False,False,False,False,False,False,True,False,False,False
+REG_TEXT_HOUSES_CALIFORNIA_PRICES_2020,True,True,False,False,False,False,True,False,False,False
+REG_TEXT_HOUSES_SAN_FRANCISCO_PERMITS_APPLICATIONS,False,False,True,True,False,False,False,False,True,False
+REG_TEXT_PROFESSIONAL_COMPANY_EMPLOYEES_SIZE,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_PROFESSIONAL_EMPLOYEE_RENUMERATION_VANCOUBER,True,True,True,True,True,True,True,True,True,True
+REG_TEXT_PROFESSIONAL_EMPLOYEE_SALARY_MONTGOMERY,False,False,True,True,True,False,True,False,True,True
+REG_TEXT_PROFESSIONAL_ML_DS_AI_JOBS_SALARIES,False,True,True,False,False,False,False,True,False,False
+REG_TEXT_PROFESSIONAL_SCIMAGOJR_ACADEMIC_IMPACT,True,True,True,True,True,True,True,True,True,True
+REG_TEXT_SOCIAL_ANIME_PLANET_RATING,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_SOCIAL_BOOKS_GOODREADS,False,True,False,False,False,False,True,False,False,False
+REG_TEXT_SOCIAL_BOOK_READABILITY_CLEAR,True,True,True,True,True,True,True,False,True,True
+REG_TEXT_SOCIAL_FILMTV_MOVIE_RATING_ITALY,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_SOCIAL_KOREAN_DRAMA,True,False,False,False,False,True,False,False,False,True
+REG_TEXT_SOCIAL_MOVIES_DATASET_REVENUE,False,False,False,False,False,False,False,False,False,True
+REG_TEXT_SOCIAL_MOVIES_ROTTEN_TOMATOES,True,True,False,True,True,True,True,True,True,True
+REG_TEXT_SOCIAL_MUSEUMS_US_REVENUES,True,False,False,False,False,True,False,False,False,True
+REG_TEXT_SOCIAL_VIDEO_GAMES_SALES,True,True,True,True,True,True,False,True,False,True
+REG_TEXT_SPORTS_FIFA22_WAGES,True,False,False,True,False,False,False,True,True,True
+REG_TEXT_SPORTS_NBA_DRAFT_VALUE_OVER_REPLACEMENT,False,False,False,False,False,False,False,False,False,False
+REG_TEXT_TRANSPORTATION_USED_CAR_MERCEDES_BENZ_ITALY,False,False,False,False,False,False,False,True,True,True
+REG_TEXT_TRANSPORTATION_USED_CAR_PAKISTAN,True,False,False,False,False,False,True,False,True,False
+REG_TEXT_TRANSPORTATION_USED_CAR_SAUDI_ARABIA,False,False,False,False,True,True,False,True,False,True
